+++ app/routes/board_routes.py
@@ -1,8 +1,7 @@
 from flask import Blueprint, request, jsonify, make_response
 from app import db
 from app.models.board import Board
-
-# from app.models.card import Card
+from app.models.card import Card
 
 # example_bp = Blueprint('example_bp', __name__)
 boards_bp = Blueprint("boards", __name__, url_prefix = "/boards")
@@ -60,6 +59,28 @@ def create_board():
     db.session.commit()
 
     return make_response(jsonify({"board": new_board.title, "owner": new_board.owner}), 201)
+
+# POST create new card
+@boards_bp.route("/<board_id>/cards", methods = ["POST"])
+def create_card(board_id):
+    # board = Board.query.get(board_id)
+    request_body = request.get_json()
+    
+    if "message" not in request_body:
+        return jsonify({"message": "Message cannot be blank!"}, 400)
+    if len(request_body["message"]) > 40:
+        return jsonify({"message": "Maximum length 40 characters."}, 400)
+
+    new_card = Card(
+        board_id = board_id,
+        message = request_body["message"],
+        likes_count = 0
+    )
+
+    db.session.add(new_card)
+    db.session.commit()
+    
+    return make_response(jsonify({"card": new_card.message}), 201)
 
 # DELETE route
 @boards_bp.route("/<board_id>", methods = ["DELETE"])
